Marker Persistence & Location Feature Enabled
All setup for including pictures has been done. Pictures will be stored
in the format '~/Adventure-UUID/num.png' where Adventure-UUID is the
adventure's unique identifier and num is the number of picture taken on
that adventure.

This is implemented so that if someone loops back around to a location,
the pictures are properly loaded.

TODO:
-Add picture functionality

-Add gridview to display pictures (likely w/ grid adapter, shouldn't
have to do a custom one)

-Add activity to view pictures once one is selected in grid view

-DONE!